# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenLearning API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/materials/{session_id}": {
      "get": {
        "summary": "Get Materials",
        "description": "Retrieve generated learning materials for a given assessment session.",
        "operationId": "get_materials_api_materials__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaterialsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MaterialOut": {
        "properties": {
          "conceptId": {
            "type": "string",
            "title": "Conceptid"
          },
          "domain": {
            "type": "string",
            "title": "Domain"
          },
          "bloomScore": {
            "type": "number",
            "title": "Bloomscore"
          },
          "qualityScore": {
            "type": "number",
            "title": "Qualityscore"
          },
          "iterationCount": {
            "type": "integer",
            "title": "Iterationcount"
          },
          "qualityFlag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qualityflag"
          },
          "material": {
            "additionalProperties": true,
            "type": "object",
            "title": "Material"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Generatedat"
          }
        },
        "type": "object",
        "required": [
          "conceptId",
          "domain",
          "bloomScore",
          "qualityScore",
          "iterationCount",
          "material",
          "generatedAt"
        ],
        "title": "MaterialOut"
      },
      "MaterialsResponse": {
        "properties": {
          "sessionId": {
            "type": "string",
            "title": "Sessionid"
          },
          "materials": {
            "items": {
              "$ref": "#/components/schemas/MaterialOut"
            },
            "type": "array",
            "title": "Materials"
          }
        },
        "type": "object",
        "required": [
          "sessionId",
          "materials"
        ],
        "title": "MaterialsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}